chore: rename lib/deskflow/clipboard_types => lib/deskflow/ClipboardTypes
