+++ src/lib/server/Server.h
@@ -14,11 +14,11 @@
 #include "common/stdset.h"
 #include "common/stdvector.h"
 #include "deskflow/Clipboard.h"
+#include "deskflow/ClipboardTypes.h"
 #include "deskflow/DragInformation.h"
 #include "deskflow/INode.h"
 #include "deskflow/KeyTypes.h"
 #include "deskflow/ServerArgs.h"
-#include "deskflow/clipboard_types.h"
 #include "deskflow/mouse_types.h"
 #include "deskflow/option_types.h"
 #include "server/Config.h"
